rename segment_ops.ts
