+++ tfjs-core/src/ops/ops.ts
@@ -58,6 +58,7 @@ export {expandDims} from './expand_dims';
 export {eye} from './eye';
 export {fill} from './fill';
 export {floorDiv} from './floorDiv';
+export {gather} from './gather';
 export {greater} from './greater';
 export {greaterEqual} from './greater_equal';
 export {imag} from './imag';
@@ -136,7 +137,6 @@ export * from './tensor_ops';
 export * from './transpose';
 export * from './softmax';
 export * from './norm';
-export * from './segment_ops';
 export * from './moving_average';
 export * from './strided_slice';
 export * from './topk';
